feat: logout with token imlemented
The logout functionality has been implemented using a token. This method uses HTTP headers to pass the token to the API.

To execute this POST request in Postman, follow these steps:

Open the POST collection.
In the Headers section, add the following parameters:
  - Authentication
  - Bearer <TOKEN>
Make sure to replace <TOKEN> with the actual token for authentication.